added logic for each read type (param vs. direct)
